+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to the "avr-c-project-generator" extension will be documente
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.2.0] - 2025-12-26
+
+### Added
+
+- Timer Calculator
+
+### Changed
+
+- Hide Clean-, Build-, and Flash-Button if no Atmel-Project is in the Workspace.
+
+## [0.1.0] - 2025-12-16
 
 - Initial release